+++ packages/lib/src/components/PayPal/Paypal.test.ts
@@ -1,11 +1,12 @@
 import Paypal from './Paypal';
+import { NO_CHECKOUT_ATTEMPT_ID } from '../../core/Analytics/constants';
 
 describe('Paypal', () => {
     test('Returns a data object', () => {
         const paypal = new Paypal({});
         expect(paypal.data).toEqual({
             clientStateDataIndicator: true,
-            paymentMethod: { subtype: 'sdk', type: 'paypal', userAction: 'pay', checkoutAttemptId: 'do-not-track' }
+            paymentMethod: { subtype: 'sdk', type: 'paypal', userAction: 'pay', checkoutAttemptId: NO_CHECKOUT_ATTEMPT_ID }
         });
     });
 
@@ -13,7 +14,7 @@ describe('Paypal', () => {
         const paypal = new Paypal({ isExpress: true });
         expect(paypal.data).toEqual({
             clientStateDataIndicator: true,
-            paymentMethod: { subtype: 'express', type: 'paypal', userAction: 'pay', checkoutAttemptId: 'do-not-track' }
+            paymentMethod: { subtype: 'express', type: 'paypal', userAction: 'pay', checkoutAttemptId: NO_CHECKOUT_ATTEMPT_ID }
         });
     });
 
@@ -21,7 +22,7 @@ describe('Paypal', () => {
         const paypal = new Paypal({});
         expect(paypal.data).toEqual({
             clientStateDataIndicator: true,
-            paymentMethod: { subtype: 'sdk', type: 'paypal', userAction: 'pay', checkoutAttemptId: 'do-not-track' }
+            paymentMethod: { subtype: 'sdk', type: 'paypal', userAction: 'pay', checkoutAttemptId: NO_CHECKOUT_ATTEMPT_ID }
         });
     });
 
@@ -29,7 +30,7 @@ describe('Paypal', () => {
         const paypal = new Paypal({ isExpress: true, userAction: 'continue' });
         expect(paypal.data).toEqual({
             clientStateDataIndicator: true,
-            paymentMethod: { subtype: 'express', type: 'paypal', userAction: 'continue', checkoutAttemptId: 'do-not-track' }
+            paymentMethod: { subtype: 'express', type: 'paypal', userAction: 'continue', checkoutAttemptId: NO_CHECKOUT_ATTEMPT_ID }
         });
     });